fix(cayenne): surface the real writer-task error instead of a generic placeholder
The partitioned Cayenne append/overwrite coordinators masked a writer
task's actual error (write failure, resource exhaustion, panic) behind
a generic "partition writer task terminated before stream end" message
whenever a sibling partition's channel closed early, discarding the
real cause before it could be logged (issue #12507).

Track why the fan-out loop stopped via a typed `WriteFanoutFailure`,
tag poison-pill echoes sent to sibling channels so join-side error
selection can tell them apart from a writer's own genuine error, and
prefer the genuine error when resolving what to report.

Also split partitioned_insert_strategy.rs into a module directory
(mod.rs/insert.rs/append.rs) and de-duplicate the partition-provider
lookup and poison-pill propagation shared by both sinks.
